+++ src/Globals/TEThemeContext/TEThemeContext.js
@@ -35,7 +35,6 @@ const initialTheme = {
 //DatePicker
 // - At this moment there is no other way to style the wrapper components on this input.
 //
-//TODO: Check if this receives props of any kind. otherwise combine baseTheme and user input theme here and then use itin this object
 const GlobalStyles = createGlobalStyle`
 	& * {
 		box-sizing: border-box;
@@ -1023,6 +1022,11 @@ const GlobalStyles = createGlobalStyle`
 	.react-datepicker__input-container input:hover {
 		border: 1px solid ${(props) => props.theme.primary};
 		box-shadow: 0 0 0 1px ${(props) => props.theme.primary} inset;
+	}
+	.react-datepicker__input-container input:disabled {
+		cursor: 'not-allowed';
+		color: ${(props) => props.theme.darkerGray};
+		border: 1px solid ${(props) => props.theme.lighterGray};
 	}
 	.react-datepicker__day--keyboard-selected {
 		background-color: ${(props) => props.theme.primary};